\name{toybox-package}
\alias{toybox-package}
\alias{toybox}
\docType{package}
\title{
Games and Simulations for Teaching Wildlife Stats
}
\description{
A set of on-screen games and simulations to take the place of the physical activities we use in our workshops if the equipment is not available.
}
\details{


So far only one function available: \code{\link{SpringPeeperCalls}}.
}



\author{
Mike Meredith

Maintainer: Mike Meredith <mmeredith@wcs.org>
}

\keyword{ package }
